Project skeleton. makefile works
In makefile, PROJECT decides what the names of the project files are
generated. For example, it generates cpu5_core.qpf and cpu5_core.qsf
when the PROJECT is cpu5_core.
ASSIGNMENT_FILES is pretty much useless, better align it with PROJECT.

When providing source file namew for quartus_map, it should be in a
form like "--source xx1.v --source xx2.v".

So I add a wrapper around SOURCE_FILES, that is
SOURCE_FILES_PARAMETER, and sent it to quartus_map.

$make
or for only comiling
$make map

compile.sh also works. Just put it here for later references.